refactor: supabase 로직을 한 곳으로 이동
