+++ fourtoon-cookie/src/components/auth/GoogleSignInAndSignUpButton.tsx
@@ -3,7 +3,7 @@ import {
     GoogleSigninButton,
 } from '@react-native-google-signin/google-signin'
 import {CLIENT_ID} from '@env'
-import {supabaseSignInAndSignUpWithIdToken} from "../../apis/supabaseSignInAndSignUpWithIdToken";
+import {supabaseSignInAndSignUpWithIdToken} from "../../apis/supabase";
 import type {Session} from "../../types/session";
 import {OAuthProvider} from "../../types/OAuthProvider";
